+++ ft_strstr.c
@@ -3,10 +3,10 @@
 /*                                                              /             */
 /*   ft_strstr.c                                      .::    .:/ .      .::   */
 /*                                                 +:+:+   +:    +:  +:+:+    */
-/*   By: epoggio <marvin@le-101.fr>                 +:+   +:    +:    +:+     */
+/*   By: epoggio <epoggio@student.42.fr>            +:+   +:    +:    +:+     */
 /*                                                 #+#   #+    #+    #+#      */
 /*   Created: 2018/10/11 18:24:19 by epoggio      #+#   ##    ##    #+#       */
-/*   Updated: 2018/10/11 19:57:15 by epoggio     ###    #+. /#+    ###.fr     */
+/*   Updated: 2018/10/29 10:55:55 by epoggio     ###    #+. /#+    ###.fr     */
 /*                                                         /                  */
 /*                                                        /                   */
 /* ************************************************************************** */
@@ -15,7 +15,7 @@
 
 char	*ft_strstr(const char *haystack, const char *needle)
 {
-	int len;
+	size_t len;
 
 	if (*needle == '\0')
 		return ((char *)haystack);
